+++ scripts/rollup/plugins/closure-plugin.js
